backend/requirements.txt
# WhatsApp 智能客户系统 - 后端依赖
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
# apscheduler==3.10.4  # 当前未使用，scheduler_service.py 使用自定义实现
python-dotenv==1.0.0
httpx==0.26.0
aiofiles==23.2.1
qrcode[pil]==7.4.2
pillow==10.2.0
openai==1.12.0
cryptography==42.0.0
neonize
pdfplumber>=0.10.0
python-docx>=1.1.0
